Re #307 Fixed MSVC warnings about int/double to tdble conversions
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3504 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 4772aa484531f33b4cbc4edc326e35418b2f0b14
Former-commit-id: 9c794edf4bd2c8d510ecca41d0dd8902e88a56ff

+++ src/libs/raceengine/raceupdate.cpp
@@ -246,7 +246,7 @@ int reSituationUpdaterThreadLoop(void *pUpdater)
 int reSituationUpdater::threadLoop()
 {
 	// Wait delay for each loop, from bRunning value (index 0 = false, 1 = true).
-	static const double KWaitDelayMS[2] = { 1.0, RCM_MAX_DT_SIMU * 1000 / 10 };
+	static const unsigned KWaitDelayMS[2] = { 1, (unsigned)(RCM_MAX_DT_SIMU * 1000 / 10) };
 
 	// Termination flag.
 	bool bEnd = false;
